requirements.txt
langchain==0.3.2
langchain-core==0.3.9
langchain-nvidia-ai-endpoints==0.3.0
langchain-text-splitters==0.3.0
langsmith==0.1.131
markdown-it-py==3.0.0
openai==1.51.0
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.39.0

